requirements.txt
Django==5.0
djangorestframework==3.14.0
pytest==7.4.3
